==> phewas/v2_imp_nc/1_phewas_loci_def.ipynb <==
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(data.table))


data_dir <- '@@@@@@/projects/biomarkers/phewas/v2_imp_nc'

hits <- '@@@@@@/users/ytanigaw/repos/rivas-lab/biomarkers/fine_mapping/filtered_hits/filtered.hits.99.anno.joined.tsv'


df <- fread(hits)

df %>% colnames()

df %>% count(Consequence)

df_coding <- df %>% filter(
    ! Consequence %in% c('missense_variant', 'stop_gained')
) 

df_coding %>%
fwrite(file.path(data_dir, 'finemapped.hits.tsv'), sep='\t')


file.path(data_dir, 'finemapped.hits.tsv')

df_coding %>% dim()

df_coding %>% 
select(chromosome, position, rsid, allele1, allele2) %>%
unique() %>% dim()

df_coding %>% 
mutate(ID = paste(chromosome, position, allele1, allele2, sep=':')) %>%
unique() %>% 
arrange(chromosome, position) %>%
select(ID) %>%
fwrite(file.path(data_dir, 'finemapped.hits.loci.txt'), sep='\t', col.names=F, row.names=F)

file.path(data_dir, 'finemapped.hits.loci.txt')
