==> phewas/v2_imp_nc/4_filter_phewas.ipynb <==
suppressWarnings(suppressPackageStartupMessages({
    library(tidyverse)
    library(data.table)
}))


res_dir <- '@@@@@@/projects/biomarkers/phewas/v2_imp_nc'
phewas_all  <- file.path(res_dir, 'ukb24983_imp_v3.glm.logistic.hybrid.gz')
phewas_hits <- file.path(res_dir, 'ukb24983_imp_v3.nc.phewas.hits.tsv')


phewas_df <- fread(cmd=paste('zcat', phewas_all)) %>% 
rename('CHROM' = '#CHROM')


phewas_df %>% select(GBE_ID) %>% unique() %>% pull() %>% length()

phewas_df %>% 
filter(
    (CHROM != 6) |
    (CHROM == 6 & POS < 25477797) | 
    (CHROM == 6 & 36448354 < POS)
) %>%
select(ID) %>% unique() %>% pull() %>% length()

0.05 / (172 * 2442)

phewas_df %>%
filter(ALT != A1) %>%
nrow()

phewas_df %>%
count(TEST, ERRCODE)


phewas_hits_df <- phewas_df %>% filter(
    P < 1e-7,
    (CHROM != 6) |
    (CHROM == 6 & POS < 25477797) | 
    (CHROM == 6 & 36448354 < POS)    
)  %>%
select(-REF, -ALT, -A1, -'FIRTH?', -TEST, -ERRCODE)


phewas_hits_df %>% dim()

phewas_hits_df %>% fwrite(phewas_hits, sep='\t')


phewas_hits
