==> phewas/v2_imp_nc/10_finngen_R2_summary.ipynb <==
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(data.table))


res_dir <- '@@@@@@/projects/biomarkers/phewas/v2_imp_nc'

phewas_hits_ld <- file.path(res_dir, 'ukb24983_imp_v3.nc.phewas.hits.ld.tsv')

finngen_R2_extracted <- file.path(res_dir, 'ukb24983_imp_v3.nc.phewas.hits.ld.finngen.R2.tsv') %>%
fread(sep='\t') %>%
rename('chrom' = '#chrom')

phewas_hits_ld_gwas_catalog_finngen_summary <- 
file.path(res_dir, 'ukb24983_imp_v3.nc.phewas.hits.ld.gwascatalog.finngen.summary.tsv')


GBE_Finngen_pheno_match <- file.path(res_dir, 'GBE_Finngen_hits.tsv')

finngen_R2_manifest <- '@@@@@@@@/users/ytanigaw/20200114_FinnGen_R2/summary_stats/r2_manifest.tsv' %>%
fread(sep='\t')


finngen_R2_df <- finngen_R2_extracted %>% 
mutate(var_id = paste(chrom, pos, ref, alt, sep='-')) %>%
left_join(
    finngen_R2_manifest  %>%
    select(phenocode, name), 
    by='phenocode'
) %>%
select(-chrom, -pos, -ref, -alt)

finngen_R2_df.colnames <- colnames(finngen_R2_df) %>%
lapply(function(x){paste0('finngen_', x)})

colnames(finngen_R2_df) <- finngen_R2_df.colnames

df <- fread(phewas_hits_ld) %>% rename('CHROM' = '#CHROM')


df %>% 
select(ID, GBE_ID, GBE_short_name, finngen_var_id) %>% 
full_join(
    finngen_R2_df %>% 
    select(finngen_var_id, finngen_name, finngen_phenocode),
    by = "finngen_var_id"
) %>% 
drop_na(ID, finngen_phenocode) %>%
select(-ID, -finngen_var_id) %>%
arrange(GBE_ID, finngen_phenocode) %>%
rename('#GBE_ID' = 'GBE_ID') %>%
unique() %>%
mutate(
    is_hit=FALSE,
    Note=''
)%>%
fwrite(GBE_Finngen_pheno_match, sep='\t')


GBE_Finngen_pheno_match

#### we performed manual annotation for the phenotypes

library(googlesheets)
gs_auth(token = "/home/users/ytanigaw/.googlesheets_token.rds")

GBE_Finngen_pheno_match_df <- 
'https://docs.google.com/spreadsheets/d/1yocReg2dL84x1NQDEyc9ywEDJM6gdfr2ELNlCcP3zPo' %>% 
gs_url() %>% 
gs_read(ws = 'GBE_Finngen_hits')


finngen_hits <- df %>%
drop_na(finngen_var_id) %>%
filter(finngen_var_id != '') %>%
unique() %>%
mutate(sort_order = 1:n()) %>%
full_join(
    finngen_R2_df,
    by='finngen_var_id'
) %>%
drop_na(ID, finngen_phenocode) %>%
left_join(
    GBE_Finngen_pheno_match_df %>%
    select(GBE_ID, finngen_phenocode, is_hit),
    by=c('GBE_ID', 'finngen_phenocode')
)


finngen_hits_summary <- finngen_hits %>%
filter(is_hit) %>%
unique() %>%
group_by(ID, GBE_ID) %>%
arrange(finngen_pval) %>%
filter(row_number() == 1) %>%
ungroup() %>%
arrange(sort_order) %>%
select(-sort_order)


phewas_hits_ld_gwas_catalog_summary_df <- 
'https://docs.google.com/spreadsheets/d/1yocReg2dL84x1NQDEyc9ywEDJM6gdfr2ELNlCcP3zPo' %>% 
gs_url() %>% 
gs_read(ws = 'ukb24983_imp_v3.nc.phewas.summary') %>% 
rename('CHROM' = '#CHROM')


finngen_hits_summary %>% colnames()

phewas_hits_ld_gwas_catalog_finngen_summary_df <- 
phewas_hits_ld_gwas_catalog_summary_df %>% 
left_join(
    finngen_hits_summary %>% 
    select(ID, GBE_ID, finngen_rsids) %>%
    rename('Finngen_hit_rsID' = 'finngen_rsids'),
    by=c('ID', 'GBE_ID')
)

phewas_hits_ld_gwas_catalog_summary_df         %>% dim() %>% print()
phewas_hits_ld_gwas_catalog_finngen_summary_df %>% dim() %>% print()


phewas_hits_ld_gwas_catalog_finngen_summary_df %>% 
rename('#CHROM' = 'CHROM') %>%
fwrite(phewas_hits_ld_gwas_catalog_finngen_summary, sep='\t')


phewas_hits_ld_gwas_catalog_finngen_summary

phewas_hits_ld_gwas_catalog_finngen_summary_df %>% 
count(is_novel, is.na(Finngen_hit_rsID))
